requirements.txt
openai
google-generativeai
ray
duckdb
rich
pydantic
pybind11
pytest
ruff
jinja2
numpy
pyyaml
